+++ changelog.md
@@ -4,6 +4,7 @@
  - added show hide button for users [users-password]
  - changed modal of teacher and student [modal]
  - admin cannot be edited or deleted from table [admin-edit-del]
+ - added password encryption for Login [password-encryption]
 
  
  # Fixed